Fixed testing for socket events
Now have connect and disconnect tests working. The issue was the client
was already connected by the time the test waiting for a 'connect' event
was executed (thanks to waiting on the connection before running tests).
So the solution was to disconnect and reconnect the client in the the
connection test cases to fire off the connect and disconnect events.