Corrected event handling errors
Now passing all connection event tests.

The issue was when I was waiting on a second client for a 'connect' or
'disconnect' error to be recieved. This was an issue because those are
reserved events for connection and disconnection in socket.io. So when a
user connects, it automatically fires a 'connect' event back to that
socket.

If you want to send another event to other clients, it nees to be a
different event.

So there are now 'user-connected' and 'user-disconnected' events for
these actions.