Updated rooms.open to return a rooms array with the new-room event
Ok so I wanted testing visibilty on the list of rooms and users (for
another commit), so when things change on these two lists I'll be
sending an array along with the event to check they've updated
correctly.

In this case I'm sending an array of room ids back with the new-room
event.

This took some doing because finding unique values in arrays in js is
not straightforward atm. I created a separate module to handle this for
the fun of it.

Also ended up having to prune undefined values before returning, and
sort the array for testing purposes.