Added specs for room-closed events
Now we're looking at the front end for the first time in a while I
noticed that rooms we're being removed from the room-list when the user
who opened the room bailed... but not when the other user did.

This is not intended behaviour, so added a couple more specs to make
sure everyone was getting the room-closed event. Which it seems they
are.

But the test runs are inconsistent, sometimes passing, sometimes running
with multiple timeouts. In the past this has been because I didn't do a
good job of timing events, which meant that I might get lucky and have
them sent, or might miss them, seemed a bit of a crapshoot. Thought I'd
sorted all of that though... perhaps not.